+++ src/Backend/spItemsRepository.ts
@@ -5,6 +5,7 @@ import {
   CompletionList,
   FileCreateEvent,
   TextDocumentChangeEvent,
+  workspace as Workspace,
 } from "vscode";
 import { URI } from "vscode-uri";
 
@@ -19,7 +20,7 @@ import {
 } from "./spFileHandlers";
 import { getAllItems, getItemFromPosition } from "./spItemsGetters";
 import { refreshDiagnostics } from "../Providers/spLinter";
-import { refreshCfgDiagnostics } from "../Providers/cfgLinter";
+import { refreshKVDiagnostics } from "../Providers/kvLinter";
 import { updateDecorations } from "../Providers/spDecorationsProvider";
 
 export class ItemsRepository implements Disposable {
@@ -47,7 +48,7 @@ export class ItemsRepository implements Disposable {
       updateDecorations(this);
       return;
     }
-    refreshCfgDiagnostics(event.document);
+    refreshKVDiagnostics(event.document);
   }
 
   public handleNewDocument(document: TextDocument) {
@@ -56,11 +57,18 @@ export class ItemsRepository implements Disposable {
       newDocumentCallback(this, document.uri);
       return;
     }
-    refreshCfgDiagnostics(document);
+    refreshKVDiagnostics(document);
   }
 
   public handleDocumentOpening(filePath: string) {
-    newDocumentCallback(this, URI.file(filePath));
+    const uri = URI.file(filePath);
+    newDocumentCallback(this, uri);
+    const doc = Workspace.textDocuments.find(
+      (e) => e.uri.fsPath === uri.fsPath
+    );
+    if (doc !== undefined) {
+      refreshKVDiagnostics(doc);
+    }
   }
 
   public getEventCompletions(): CompletionList {